Rename subscription manager and worker from plural to singular
