+++ apps/subscription-manager/src/subscription_manager.py
